Initial bigtable/api example with automated builds. (#2)
* Create skeleton grpc program.

* Initial API example.

Kind of trivial, but need to start somewhere, list all instances.

* Add comments in list_instances program.

The point of these examples is for users to understand what is
going on.  An excess of comments cannot hurt in that case.

* Initial configuration to integrate with Travis CI.

* Test with g++ only

* Remove clang package because it breaks build.

* Use sudo to install protobuf and gRPC++.

* Delete obsolete files.

* Enable sudo, too hard to install packages otherwise.

* Cache grpc build directory and install in $HOME.

* Try to fix cached grpc builds.

* Print the full environment.

* Print the full environment.

* Pick one name for grpc-build and stick to it.

* Another attempt at enabling sudo.

* Join update-alternatives command to one line.

Seems the backslash at end of line confused things.

* Only require version 3.2 for cmake.

Ubunut 14.04 (Trusty), the preferred platform on Travis, comes
with an ancient cmake version, so use that.

* Add C as one of the languages for the project.

Evidently cmake 3.7 can detect pthreads using C++, but 3.2 needs
C enabled.

* Remove remnants of builds in bigtable/ directory.

* Generate the googleapis protos before compiling samples.

* Use clang instead of gcc as the default compiler.

On the Travis image g++ seems to be too old, go with clang and
see if that works.

* Require C++11 from cmake.

Evidently I do not know how to require C++11 from cmake, I think
I got it for both newer and older versions now.

* Add debugging for installed clang version.

For some reason the script is picking up clang-3.5 when I install
clang-3.9 explicitly.  Try to debug that.

* More debugging of clang++ installation.

* In cmake, use value of CXX to pick compiler.

Strangely clang++ is in the path and it is the right version, but
cmake prefers something installed in /usr/local/blahblah.

* Stash docker builds in a branch.

* Enable g++ build and use full path of compiler in cmake.

* Cache each grpc build separately, fix typos.

* Arggh, fix another typo in travis script.

* Fix warning on clang++ build.

* Fix more syntax errors in build script.

* Use TRAVIS settings for clone and checkout.

* Prepare for caching docker images.

* Fix bad tabs.

* Fixed more syntax errors in build script.

* Several fixes to work on Fedora 24.

* Enable one docker-based build.

* Fix syntax error in docker-based build.

* Use the C++ compiler provided as a build argument.

* Set the CXX_COMPILER flag in the build matrix.

* Remove spurious \ in docker-based build script.

* Parallelize the build of the C++ examples.

Use 2 CPUs for the moment, that is what Travis has anyways.
We could try to determine the number of available CPUs and
maximize the parallelism, but meh...

* Only use --cache-from if the image was available.

* Another syntax error, really need local tests.

* More trivial fixes.

* Only build under docker.

The cache does not work for non-docker builds, and it is so hard
to configure the environment as to make it useless.

* Add debugging for docker image caching.

Something is not working as I expect.

* Seems like I need to update docker in the Travis build.

* Cleanup debugging statements.

Remove some that are not needed, also sort others and minimize directories
that are cached.

* Copy travis-prepared working copy to the image.

This simplifies creating the same working copy of branches, regular
commits, pull-requests, etc.

* Removed too much debugging in previous commit.

* Further simplification of the docker configuration.

* Fix running directory for docker build.

* Enable two more builds.

This should also test the caching after a build.

* Save images before the cache is uploaded.

* Do not install ca-cacert because Ubuntu:14.04 lacks it.

* Always install the default clang/clang++ compilers.

* Give up on Ubuntu:14.04 tools are too old.

* Use variables for more readable scripts.

* Enable all builds.

* Address comments in pr/18.