Implement MutateRow() and MutateRows() examples. (#5)
* Create skeleton grpc program.

* Initial API example.

Kind of trivial, but need to start somewhere, list all instances.

* Add comments in list_instances program.

The point of these examples is for users to understand what is
going on.  An excess of comments cannot hurt in that case.

* Initial configuration to integrate with Travis CI.

* Test with g++ only

* Remove clang package because it breaks build.

* Use sudo to install protobuf and gRPC++.

* Delete obsolete files.

* Enable sudo, too hard to install packages otherwise.

* Cache grpc build directory and install in $HOME.

* Try to fix cached grpc builds.

* Print the full environment.

* Print the full environment.

* Pick one name for grpc-build and stick to it.

* Another attempt at enabling sudo.

* Join update-alternatives command to one line.

Seems the backslash at end of line confused things.

* Only require version 3.2 for cmake.

Ubunut 14.04 (Trusty), the preferred platform on Travis, comes
with an ancient cmake version, so use that.

* Add C as one of the languages for the project.

Evidently cmake 3.7 can detect pthreads using C++, but 3.2 needs
C enabled.

* Remove remnants of builds in bigtable/ directory.

* Generate the googleapis protos before compiling samples.

* Use clang instead of gcc as the default compiler.

On the Travis image g++ seems to be too old, go with clang and
see if that works.

* Require C++11 from cmake.

Evidently I do not know how to require C++11 from cmake, I think
I got it for both newer and older versions now.

* Add debugging for installed clang version.

For some reason the script is picking up clang-3.5 when I install
clang-3.9 explicitly.  Try to debug that.

* More debugging of clang++ installation.

* In cmake, use value of CXX to pick compiler.

Strangely clang++ is in the path and it is the right version, but
cmake prefers something installed in /usr/local/blahblah.

* Stash docker builds in a branch.

* Enable g++ build and use full path of compiler in cmake.

* Cache each grpc build separately, fix typos.

* Arggh, fix another typo in travis script.

* Fix warning on clang++ build.

* Fix more syntax errors in build script.

* Use TRAVIS settings for clone and checkout.

* Prepare for caching docker images.

* Fix bad tabs.

* Fixed more syntax errors in build script.

* Several fixes to work on Fedora 24.

* Enable one docker-based build.

* Fix syntax error in docker-based build.

* Use the C++ compiler provided as a build argument.

* Set the CXX_COMPILER flag in the build matrix.

* Remove spurious \ in docker-based build script.

* Parallelize the build of the C++ examples.

Use 2 CPUs for the moment, that is what Travis has anyways.
We could try to determine the number of available CPUs and
maximize the parallelism, but meh...

* Only use --cache-from if the image was available.

* Another syntax error, really need local tests.

* More trivial fixes.

* Only build under docker.

The cache does not work for non-docker builds, and it is so hard
to configure the environment as to make it useless.

* Add debugging for docker image caching.

Something is not working as I expect.

* Seems like I need to update docker in the Travis build.

* Cleanup debugging statements.

Remove some that are not needed, also sort others and minimize directories
that are cached.

* Copy travis-prepared working copy to the image.

This simplifies creating the same working copy of branches, regular
commits, pull-requests, etc.

* Removed too much debugging in previous commit.

* Further simplification of the docker configuration.

* Fix running directory for docker build.

* Enable two more builds.

This should also test the caching after a build.

* Save images before the cache is uploaded.

* Do not install ca-cacert because Ubuntu:14.04 lacks it.

* Always install the default clang/clang++ compilers.

* Give up on Ubuntu:14.04 tools are too old.

* Use variables for more readable scripts.

* Enable all builds.

* Use *_LIBRARY_DIRS and similar for grpc and friends.

pkg_check_modules() sets a number of useful variables to add -I and
-L and -l flags, and I just forgot to use them.

* Refactor Travis install and build scripts.

The .travis.yml file was getting large and hard to follow, but
the main motivation was to make it easier to debug the build
scripts on a local machine before pushing to Travis.

* Simplify build script.

I am assuming that brew install works correctly, and sets up all
the necessary links in /usr/local/bin and /usr/local/lib.  I would
rather make the install script more complex than the build script.

* Fix name of OS.

* Separate googleapis build for easier debugging.

Running on a separate stage makes the build go faster the second
time.

* Handle errors in `docker inspect`.

Do not abort the script if `docker inspect` fails, simply continue
without a previous image serving as cache.

* Rationalize Linux builds.

Instead of full coverage of all compiler + OS combos only build
with the combos that are likely to catch interesting failures.

* Use CXX and CC environment variables to pick compiler.

Travis already sets both variables, there is no need to create a
separate variable.  Well, not in the current setup where the
compiler can be found via PATH, if we wanted to use a compiler not
in the PATH we would need to.

* New example to create an instance.x

* New example showing how to delete an instance.

* Fix formatting errors.

* I think this is the right order for header files.

* Fix formatting.

* A new example on how to create a table.

* Use the (undocumented) result type for CreateInstance().

The CreateInstance() long running operation returns a Instance
when successful.  There is an outstanding bug to fix the
documentation.  We can use the correct type in the example to
produce a better result message.

* Fixes to compile on Windows with MSVC 2015.

* Initial scripts to automate builds on Windows.

* Initial configuration for AppVeyor.

* Call powershell explicitly to invoke scripts.

* Debug appveyor script.

* Fix name of install script and try again.

* Cache gRPC build output.

* Trying to minimize the cache size.

* Fix syntax error and install something smaller.

The smaller install will speed up the test cycles.

* Debug caching directives.

* Fix dependencies for cache directory.

* Fix install when vcpkg\installed is cached.

* Build grpc and its dependencies.

* Compile both x86 and x64 versions.

* Multiple fixes to compile with minimal warnings.

I had to use static libraries because protobufs emit inconsistent
dllimport/dllexport declarations for some common headers.
I had to disable most warnings, because protobufs have massive
numbers of them too.

* Actually install static libraries.

I tested everything with static libraries locally, and then committed
the version with dynamic libraries, grr....

* Fix debugging statements.

* Always save cache results.

We need it to make progress when builds fail during install or
timeout during install.

* Enable build script.

* Add build status badges.

* Use the TAQ proto to store the data in Bigtable.

* Uncompress the file after downloading it.

* Write a demo for the MutateRows() API.

* Fixed bugs in mutate_with_retries().