feat: index GCS metadata in Cloud Spanner
In this PR we introduce a new example that can index GCS metadata (the
names and properties of all the objects in a Bucket) using Cloud
Spanner. One of the programs in this example can be deployed to Cloud
Run to maintain the index as objects are modified, removed, or uploaded
to GCS. There are also examples showing how to run the program to index
a bucket on Google Kubernetes Engine (GKE).

The example uses the (existing) C++ client libraries for Google Cloud,
and showcases how to deploy C++ applications to GCP.