doc: another iteration on GKE example (#165)
In this iteration I completed the documentation to create the GKE
cluster, deploy the application to said cluster, and run some test data
through the application. GKE needs a lot more configuration, and some
advanced use-cases (scaling up automatically) are beyond what we should
cover in a "Getting Started with C++" guide.