fix: get the speech/ example working on Windows
I think the specific problem that made this example fail affected all
platforms, namely, that it read 0 bytes from a file (this can happen at
EOF if the file is a multiple of the read buffer size), and then tried
to send this 0-byte blob to the service.

To get the code working on Windows I had to replace `getopt(1)` with
something more portable, I chose Boost::program_options because we use
it in other places.

I also made a pass making the code look more idiomatic for C++11: prefer
using range-based for loops, prefer returning values, use exceptions
(instead of `nullptr`) to signal errors, etc.

We should make a second pass fixing this code to use the C++ client
library, calling gRPC directly is exceedingly painful.