ci: switch builds to top-level CMake file
Maintaining the YAML file for Google Cloud Build was getting tedious.
Using a top-level CMake file would be easier to maintain, all we would
need to do is add new directories to the sample list.  The top-level
CMake file is a super build. This is to validate the configure+build
steps in each directory work. In contrast, using `add_subdirectory()`
would build everthing, but would configure only once, using the
top-level CMake file.